requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
openai==1.69.0
python-dotenv==1.0.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
pydantic==2.0.3